✨ feat: 更新damage list, 现在call多个damage的技能的伤害会被如游戏内合并

+++ source/src/modules/task.cpp
@@ -437,7 +437,7 @@ std::string Task::queryDamageList() {
             }
             json objDamage;
             objDamage["time"]           = everyDamage.tick / 1024.0;
-            objDamage["type"]           = static_cast<int>(everyDamage.damageType);
+            objDamage["type"]           = everyDamage.damageType;
             objDamage["id"]             = everyDamage.id;
             objDamage["level"]          = everyDamage.level;
             objDamage["damageBase"]     = everyDamage.damageBase;
